A console calculator in 16bits ASM with TASM
This program is 16 bits calculator in console programming with TASM that
allow you to do addition , subtraction , division for integer without
parentheses. In the futur we will make it interprets parentheses.

To do that :
1)  analyse the expression
2 ) Convert the INFIXE string to a postfixe string
3) we evaluate the Postfixe string

You will show also a procedure that allow you calculate the factorial of
a number from 0 to 8.